PermissionDenied for unregister user

+++ MainApp/views.py
@@ -3,6 +3,8 @@ from django.shortcuts import render, redirect
 from MainApp.models import Snippet
 from MainApp.forms import SnippetForm, UserRegistrationForm, CommentForm
 from django.contrib import auth
+from django.contrib.auth.decorators import login_required
+from django.core.exceptions import PermissionDenied
 
 
 def index_page(request):
@@ -64,8 +66,11 @@ def snippet_detail(request, snippet_id):
     return render(request, 'pages/snippet_detail.html', context)
 
 
+@login_required
 def snippet_delete(request, snippet_id):
     snippet = Snippet.objects.get(id=snippet_id)
+    if snippet.user != request.user:
+        raise PermissionDenied()
     snippet.delete()
     return redirect("snippets-list")
